requirements.txt
# Dépendances principales
psycopg2-binary==2.9.9  # Connecteur PostgreSQL
redis==5.0.1            # Client Redis
python-dotenv==1.0.1    # Gestion des variables d'environnement

# Dépendances du projet SEED Tech
numpy==2.0.2
openai==1.58.1
Flask==3.0.0
waitress==3.0.0
Flask-CORS==5.0.0
requests==2.32.3
beautifulsoup4==4.13.3
python-dateutil==2.9.0.post0

# Outils de développement et type checking
types-redis==5.0.1
types-python-dateutil==2.9.0.20240311